refactor(admin): apply review feedback — chart no-mvp + availability split
Follow-up to code-reviewer feedback on PR #185:

1. StudentAcademicCharts: drop placeholder builders that built 3 dots
   at null grades + 3 zero bars absences. These never rendered (empty
   states already gated them) but violated no-mvp-only-premium as code
   smell. Series default to []; the empty states 'Pas encore de notes'
   et 'Aucune absence' rendent toujours.

2. Split TeacherAvailabilityTab (450 LOC -> 204 LOC orchestrator +
   3 sub-modules) per detail-tab-split-pattern.md:
   - availability-helpers.ts   (DAYS, HOURS, types, cellKey, buildSavedMap)
   - AvailabilityGrid.tsx      (pure stateless table render)
   - AvailabilityToolbar.tsx   (read/edit toolbar + Legend)
   - tabs/TeacherAvailabilityTab.tsx (state machine + handlers, kept import path)
   All files now under the 250 LOC soft limit.
